requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
mlflow>=2.0.0
datasets>=2.0.0
pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0




